% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{is_language}
\alias{is_language}
\title{Is \code{x} a language?}
\usage{
is_language(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
\itemize{
\item \code{TRUE} if \code{x} is a \code{"tree_sitter_language"}.
\item \code{FALSE} otherwise.
}
}
\description{
Use \code{is_language()} to determine if an object has a class of
\code{"tree_sitter_language"}.
}
