% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{is_query}
\alias{is_query}
\title{Is \code{x} a query?}
\usage{
is_query(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
\code{TRUE} if \code{x} is a \code{tree_sitter_query}, otherwise \code{FALSE}.
}
\description{
Checks if \code{x} is a \code{tree_sitter_query} or not.
}
\examples{
\dontshow{if (rlang::is_installed("treesitter.r")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
source <- "(identifier) @id"
language <- treesitter.r::language()

query <- query(language, source)

is_query(query)

is_query(1)
\dontshow{\}) # examplesIf}
}
