% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language_field_count}
\alias{language_field_count}
\title{Language field count}
\usage{
language_field_count(x)
}
\arguments{
\item{x}{\verb{[tree_sitter_language]}

A tree-sitter language object.}
}
\value{
A single double value.
}
\description{
Get the number of fields contained within a language.
}
\examples{
\dontshow{if (rlang::is_installed("treesitter.r")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
language <- treesitter.r::language()
language_field_count(language)
\dontshow{\}) # examplesIf}
}
