% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language_id_for_node_kind}
\alias{language_id_for_node_kind}
\title{Language IDs}
\usage{
language_id_for_node_kind(x, kind, ..., named = TRUE)
}
\arguments{
\item{x}{\verb{[tree_sitter_language]}

A tree-sitter language object.}

\item{kind}{\verb{[character]}

The node kinds to look up the ID for.}

\item{...}{These dots are for future extensions and must be empty.}

\item{named}{\verb{[logical]}

Should named or anonymous nodes be looked up? Recycled to the
size of \code{kind}.}
}
\value{
An integer vector the same size as \code{kind} containing either:
\itemize{
\item The integer ID of the node kind, if known.
\item \code{NA} if the node kind was not known.
}
}
\description{
Get the integer language ID for a particular node kind. Can be useful
for exploring the grammar.
}
\seealso{
\code{\link[=language_node_kind_for_id]{language_node_kind_for_id()}}
}
