% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language_name}
\alias{language_name}
\title{Language name}
\usage{
language_name(x)
}
\arguments{
\item{x}{\verb{[tree_sitter_language]}

A tree-sitter language object.}
}
\description{
Extract a language object's language name.
}
