% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language_node_kind_for_id}
\alias{language_node_kind_for_id}
\title{Language node kinds}
\usage{
language_node_kind_for_id(x, id)
}
\arguments{
\item{x}{\verb{[tree_sitter_language]}

A tree-sitter language object.}

\item{id}{\verb{[positive integer]}

The language IDs to look up node kinds for.}
}
\value{
A character vector the same length as \code{id} containing:
\itemize{
\item The name of the node kind, if known.
\item \code{NA}, if the ID was not known.
}
}
\description{
Get the string node kind for a particular language ID. Can be useful for
exploring a grammar.
}
\seealso{
\code{\link[=language_id_for_node_kind]{language_id_for_node_kind()}}
}
