% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language-r.R
\name{language_r}
\alias{language_r}
\title{R tree-sitter language}
\usage{
language_r()
}
\value{
A \code{tree_sitter_language} object.
}
\description{
Generates a \code{tree_sitter_language} object for R.
}
\details{
This language object was built from the following:
\itemize{
\item URL: https://github.com/r-lib/tree-sitter-r
\item SHA: 03e6c381ba3d3d4802865e0108c56c86a6ef3f85
}
}
\examples{
language_r()
}
