% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language_version}
\alias{language_version}
\title{Language version}
\usage{
language_version(x)
}
\arguments{
\item{x}{\verb{[tree_sitter_language]}

A tree-sitter language object.}
}
\value{
A scalar double value.
}
\description{
Get the ABI version for this language object.
}
