% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{new_language}
\alias{new_language}
\title{Construct a new tree-sitter language object}
\usage{
new_language(name, pointer)
}
\arguments{
\item{name}{\verb{[string]}

The name of the language being wrapped.}

\item{pointer}{\verb{[external_pointer]}

An external pointer to a \verb{const TSLanguage*}.}
}
\value{
A \code{tree_sitter_language} object.
}
\description{
This function is a developer tool to wrap an external pointer to a
C level tree-sitter \verb{const TSLanguage*}.

The \code{language()} function exported from your grammar specific R package
should call this, providing the language name and an external pointer to
the result of the C level \verb{tree_sitter_\{name\}()} function.
}
