% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter-package.R
\docType{package}
\name{treesitter-package}
\alias{treesitter}
\alias{treesitter-package}
\title{treesitter: R Bindings to tree-sitter}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
