% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{x_tree_sitter_node}
\alias{x_tree_sitter_node}
\title{Helper page for consistent documentation}
\arguments{
\item{x}{\verb{[tree_sitter_node]}

A node.}
}
\description{
Helper page for consistent documentation
}
\keyword{internal}
