% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{x_tree_sitter_tree}
\alias{x_tree_sitter_tree}
\title{Helper page for consistent documentation}
\arguments{
\item{x}{\verb{[tree_sitter_tree]}

A tree.}
}
\description{
Helper page for consistent documentation
}
\keyword{internal}
