% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter-package.R
\docType{package}
\name{treesitter-package}
\alias{treesitter}
\alias{treesitter-package}
\title{treesitter: R Bindings to tree-sitter}
\description{
Provides bindings to tree-sitter, an incremental parsing system for programming tools. Tree-sitter builds concrete syntax trees for source files of any language, and can efficiently update those syntax trees as the source file is edited. It also includes a robust error recovery system that provides useful parse results even in the presence of syntax errors.
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item tree-sitter authors (tree-sitter C library) [copyright holder]
}

}
\keyword{internal}
