% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesitter-package.R
\docType{package}
\name{treesitter-package}
\alias{treesitter}
\alias{treesitter-package}
\title{treesitter: Bindings to 'Tree-Sitter'}
\description{
Provides bindings to 'Tree-sitter', an incremental parsing system for programming tools. 'Tree-sitter' builds concrete syntax trees for source files of any language, and can efficiently update those syntax trees as the source file is edited. It also includes a robust error recovery system that provides useful parse results even in the presence of syntax errors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/r-tree-sitter}
  \item \url{https://davisvaughan.github.io/r-tree-sitter/}
  \item Report bugs at \url{https://github.com/DavisVaughan/r-tree-sitter/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Tree-sitter authors (Tree-sitter C library) [copyright holder]
}

}
\keyword{internal}
